Starting to put together commodity lists.

+++ Source/Libraries/YpsilonExt/Ypsilon/Entities/Ship.cs
@@ -9,7 +9,7 @@ namespace Ypsilon.Entities
     {
         public AShipDefinition Definition = new AShipDefinition();
 
-        public float ResourceOre = 0f;
+        public int ResourceOre = 0;
 
         public Ship(EntityManager manager, Serial serial)
             : base(manager, serial)
